rename variables

+++ src/simd/avx.cpp
@@ -431,7 +431,7 @@ SQ4UniformComputeCodesIP(const uint8_t* codes1, const uint8_t* codes2, uint64_t 
     }
     alignas(256) int16_t temp[16];
     int32_t result = 0;
-    uint32_t d = 0;
+    uint64_t d = 0;
     __m256i sum = _mm256_setzero_si256();
     __m256i mask = _mm256_set1_epi8(0xf);
     for (; d + 63 < dim; d += 64) {
