+++ src/simd/avx512.cpp
@@ -327,7 +327,7 @@ SQ4UniformComputeCodesIP(const uint8_t* codes1, const uint8_t* codes2, uint64_t 
     }
     alignas(512) int16_t temp[32];
     int32_t result = 0;
-    uint32_t d = 0;
+    uint64_t d = 0;
     __m512i sum = _mm512_setzero_si512();
     __m512i mask = _mm512_set1_epi8(0xf);
     for (; d + 127 < dim; d += 128) {
